OID for the MIB root changed to SAF ratified. Fixed compiler warnings for Event and SEL table. Updated README file to reflect current situation

+++ src/saHpiEventTable.c
@@ -1978,7 +1978,7 @@ saHpiEventTable_get_value (netsnmp_request_info * request,
 				(char *) &context->saHpiEventDelete,
 				sizeof (context->saHpiEventDelete));
 	*/
-      return NULL;
+      return (int)NULL;
       break;
 
     default:
